Security: Use modern API_AVAILABLE and API_DEPRECATED
Apple no longer includes all the macros in `AvailabilityInternal.h` in
the open source release of XNU. We could get the file from macOS
installs because it looks like those do include the macros for all the
possible deprecated/added version combinations. However, Apple has also
introduced a different way of specifying the availability attribute,
which no longer leads to a combinatorial explosion of combinations, and
a huge header (20k+ lines). So instead of patching XNU, we can patch
places where the old-style macros are used with versions past 10.12.4.