bootstrap-tools: Add SDK unpacking deps
We need to be able to unpack the SDK earlier in the bootstrap so we can
provide XPC to the Apple source releases. So we need `pbzx` and that in
turn requires dylibs from `xar` and `bzip2`. These are already used for
the aarch64 bootstrap-tools so now we add them unconditionally.