xnu: Fix Availability for *_NA platform macros
The previous patch tried to deal optimistically with all the potential
version macros, `__MAC_xx_yy[_zz]`, `__IPHONE_xx_yy[_zz]`, but this fell
short of the mark. The old Availability macros were inflexible and
`__MAC_NA` and `__IPHONE_NA` version macros were necessary as a
workaround for cases where one of the version constraints did not have
to or could not be specified. The naive implementation of the old macros
could not deal with these because it's impossible to translate no
constraint at all to a constraint that is semantically equivalent. (I
was tempted to just add `API___MAC_NA MAX_INT.MAX_INT.MAXINT` though.)

This patch more thoroughly implements the old macros, including the
`*_NA` workaround being handled to not result in a constraint.