ICU: Unbump macOS 11.0.1 -> 10.13.6
Somehow ICU got bumped to the version that comes with 11.0.1. This is
inconsistent with all the other source releases so I'm downgrading the
version. This requires some changes to the ICU expression that will
break the build again when updating to ICU 62107.0.1 or later, starting
with macOS 10.14.

Since the source releases aren't separate for aarch64-darwin this means
a downgrade of the ICU version there as well but that will bring it in
line with the other source releases.

Commit 145ed3fcbc06a92fba6efd142a957cc6321ecb21 has the changes to the
ICU expression to make it work with newer versions of ICU.