Security: Drop Availability patch retaining subset
The patch to switch every occurrence of the old Availability macros is
no longer necessary. However, one header in the framework has typos for
the `__PHONE_NA` macros and these are key to the implementation of the
modern-approach backed old Availability macros so we still need to patch
those.

Apple has fixed this in later versions of the header.