bootstrap_cmds: Unbump macOS 11.0.1 -> 10.13.6
Like ICU, bootstrap_cmds got bumped to the version that comes with
11.0.1. This is inconsistent with all the other source releases so I'm
downgrading the version.

Since the source releases aren't separate for aarch64-darwin this means
a downgrade of the bootstrap_cmds version there as well but that will
bring it in line with the other source releases.

Commit 84410e29df3787f60d16f327e6dc43f5bcba2e4e originally removed the
package from `apple-source-releases/default.nix`.