Libsystem: Preserve symlinks in copyHierarchy
Apple uses many symlinks throughout the SDK. Copying the files as normal
files rather than symlinks isn't a problem for users of Libsystem, it
just increases disk usage a bit. However, the header check ignores
symlinks and that means header lists are inconsistent. For Libsystem it
lists everything because symlinks become regular files and are therefore
not ignored. For other frameworks header files are incomplete because
they never checked for symlinks.

By preserving symlinks in Libsystem and not excluding symlinks in the
header check this should all become more consistent.