system_cmds: Fix command builds
Switch to the from-source IOKit.

Clearly indicate whether something is being built or skipped.

Fix the builds for `dmesg`, `ltop`, `proc_uuid_policy`, `taskpolicy` and
`vm_purgeable_stat`.

Start skipping the builds for `gcore`, `lskq`, `lsmp` and `zprint`
because of unfulfillable dependencies.

Stop skipping the builds for `dirhelper` and `dynamic_pager`  and copy
`pagesize` without building because it's a shell script.